refactored into four main files
